+++ pages/human-resources/doctor/add/index.js
@@ -1,35 +1,37 @@
 import { useRouter } from "next/router";
 import { useSnackbar } from "notistack";
-import AddNurseForm from "../../../../src/components/AddNurseForm";
+import AddDoctorForm from "../../../../src/components/AddDoctorForm";
+import Section from "../../../../src/components/Section";
 import { errorNotify, successNotify } from "../../../../src/constants/notistackVariants";
-import { CREATE_NEW_NURSE_URL } from "../../../../src/constants/url";
+import { CREATE_NEW_DOCTOR_URL } from "../../../../src/constants/url";
 import { defaultPage } from "../../../../src/hocs/defaultPage";
 import { protectRoute } from "../../../../src/hocs/protectRoute";
 import { useTranslation } from "../../../../src/i18n";
 
-function AddNewNursePage() {
-
-  const { enqueueSnackbar } = useSnackbar();
+function AddNewDoctorPage(props) {
   const { t } = useTranslation();
+  const { enqueueSnackbar } = useSnackbar();
   const router = useRouter();
 
-  const handleAddNewNurse = (hisCode, name, sex, birthDate) => {
-    fetch(CREATE_NEW_NURSE_URL, {
+  const handleAddNewDoctor = (hisCode, name, sex, email, phoneNumber, birthDate) => {
+    fetch(CREATE_NEW_DOCTOR_URL, {
       method: "post",
       headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json'
+        Accept: "application/json",
+        "Content-Type": "application/json",
       },
       body: JSON.stringify({
         hisCode: hisCode,
         name: name,
         gender: sex,
-        birthDate: birthDate,
+        email: email,
+        phoneNumber: phoneNumber,
+        birthDay: birthDate,
       }),
     })
       .then((response) => {
         if (response.ok) {
-          enqueueSnackbar(t("Create nurse successful"), successNotify);
+          enqueueSnackbar(t("Create doctor successful"), successNotify);
           router.back();
           return response.json();
         } else {
@@ -40,16 +42,16 @@ function AddNewNursePage() {
         console.log(data);
       })
       .catch((error) => {
-        enqueueSnackbar(t("Error during creating nurse"), errorNotify);
+        enqueueSnackbar(t("Error during creating doctor"), errorNotify);
         console.error(error);
       });
   };
 
   return (
     <div>
-      <AddNurseForm handleAddNewNurse={handleAddNewNurse} />
+      <AddDoctorForm handleAddNewDoctor={handleAddNewDoctor} />
     </div>
   );
 }
 
-export default protectRoute(defaultPage(AddNewNursePage));
+export default protectRoute(defaultPage(AddNewDoctorPage));