+++ src/constants/url.js
@@ -1,2 +1,7 @@
-export const GET_ALL_NURSES_URL = process.env.serverUrl + "/nurses";
-export const CREATE_NEW_NURSE_URL = process.env.serverUrl + "/nurses"
+const host = process.env.serverUrl;
+
+export const GET_ALL_NURSES_URL = host + "/nurses"
+export const CREATE_NEW_NURSE_URL = host + "/nurses"
+
+export const GET_ALL_DOCTORS_URL = host + "/doctors"
+export const CREATE_NEW_DOCTOR_URL = host + "/doctors"